[3798] Changes after review:
 - packet names updated in User's Guide
 - StatsMgr now uses int64_t, not uint64_t
 - StatsMgr::generateName() added
 - DISCOVER => DHCPDISCOVER

+++ src/lib/dhcpsrv/cfgmgr.cc
@@ -228,7 +228,7 @@ CfgMgr::updateStatistics() {
         name << "subnet[" << (*subnet)->getID() << "].total-addresses";
 
         StatsMgr::instance().setValue(name.str(),
-                                      (*subnet)->getPoolCapacity(Lease::TYPE_V4));
+            static_cast<int64_t>((*subnet)->getPoolCapacity(Lease::TYPE_V4)));
     }
 }
